fix(bin): a workaround for using 'sed -i' on a MacOs terminal
On mac, unlike Linux, we when we use 'sed -i', the sed command expects a file name just after the '-i'
It turns out that passing blank ('') as in "-i ''" is not working.

For a workaround, I am passing a file extension namely .remove which we will use as a marker to erase all *.remove in the end.